feat: add a get_guild_db method
Not sure how to proceed. We can't sync database at startup, but then some guilds may not be in the database.
Also, we can't check directly the bot is still in a guild using the db, but there will be http connection between bot and dashboard, so it is not a problem.

At the moment, we consider that all GuildDB objects will be retrieved within this method. But this will maybe lead of some issues when we will want to join other db using relationships.

It is still under discussion (with myself)