feat: implement misc_commands and then SpecialCog
misc_command are a sort of commands but that works within events. For example, we can create a misc_command "translate" over the event on_raw_reaction_add.

This is usefull to set a description, etc... for this type of features.

To automate these stuffs, I made a SpecialCog class that will automatically save the misc_commands under a special attribute. It also has a default __init__ method that bind the bot to the Cog.