feat: misc command checkers are done, or almost
It was a mess.
We ignore all event error, since we consider there is no reason an error should be raised in an event that is NOT a misc_command.
Errors raised in misc commands are dispatched within the "on_misc_commands_error" event.
misc_guild_only has been created, to assert a misc command is executed in a guild.

I think misc commands are compatible with a majority of events, especially within the "on_reaction_*" (and raw as well), and "on_message" events.