feat: create support for checkers
Add a checker bot_required_permissions to check if the bot has the right permissions before executing the command. This will also store these permissions into cmd.extra (for either MiscCommand and AppCommands).

Because of AppCommands limitation, the checker needs to be above the @command() decorator (to be able to add the permissions into extra)