Add 2048 (#110)
Merge this branch to clean a little the github repo.
This game can be considered as ready, but in any case, it is marked as "beta" AND the bot is still in beta.
I can also disable the feature when running the bot. Merging the branch will reduce merge conflits for later.