[wip] add a feature exporter utility
This program will read the code, to extract all the informations related to the commands. It will then generate a json file with these informations. This will be used to generate the documentation in the future.
My work here is not guaranteed to work correctly. I just commit it, to allow someone else to continue.

By the way, all the mentions of "commands" in the filenames should be renamed to "features".