replace example files with documentation
To not bloat the codebase with unnecessary files, the informations are now in the CONTRIBUTING.md file. This also allow to add some explanations.

Some code refactor is needed to cleanup some things such a the `TRANSLATOR_SERVICES` that should be a configuration value. Some configs should be optional as well. These informations will be updated in the futur.